restaurantAPI/requirements.txt
alembic==1.11.1
asgiref==3.8.1
asttokens==2.4.0
attrs==23.2.0
autoenv==1.0.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
coverage==7.6.0
coveralls==4.0.1
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.6
Django==5.0.7
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
ecdsa==0.18.0
executing==1.2.0
filelock==3.12.2
Flask==1.1.2
Flask-API==3.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
h11==0.14.0
idna==3.6
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
outcome==1.3.0.post0
parso==0.8.3
passlib==1.7.4
pbr==5.11.1
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
sqlparse==0.5.0
stack-data==0.6.2
stevedore==5.1.0
traitlets==5.10.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
Werkzeug==2.3.6
wsproto==1.2.0
